Snap: NumMatrix.java, Solution.java

+++ ThreadTry/src/Solution.java
@@ -125,11 +125,11 @@ class Solution {
         return result;
     }
 
-    public static int[] sortByBits(Integer[] arr_list) {
+    public static int[] sortByBits(int[] arr_list) {
         Arrays.sort(arr_list, (a, b)->{
             String bin_a = int2Binary((Integer) a);
             String bin_b = int2Binary((Integer) b);
-            return (!bin_a.equals(bin_b) ? bin_a.compareTo(bin_b):Integer.compare(a, b));
+            return (!bin_a.equals(bin_b) ? bin_a.compareTo(bin_b):Integer.compare((Integer) a, (Integer) b));
         });
         return arr_list;
     }
